+++ rs/tests/src/consensus/mod.rs
@@ -1,7 +1,6 @@
 pub mod catch_up_test;
 pub mod cow_safety_test;
 pub mod liveness_with_equivocation_test;
-pub mod node_graceful_leaving_test;
 pub mod payload_builder_test;
 pub mod replica_determinism_test;
 pub mod request_auth_malicious_replica_test;
